PORT: Make janus.c using the MSVC compiler.
MSVC is the default compiler required by the pip version shipped with
the default Windows installers.

+++ janus/hash.c
@@ -94,7 +94,7 @@ static unsigned int
 MurmurHashAligned2(const void * key, size_t len, unsigned int seed)
 { const unsigned int m = 0x5bd1e995;
   const int r = 24;
-  unsigned int h = seed ^ len;
+  unsigned int h = seed ^ (unsigned int)len;
   const unsigned char * data = (const unsigned char *)key;
 
   while( len >= 4 )
